+++ app/OpenAPI/Schemas/ShoppingList.php
@@ -9,10 +9,8 @@ use OpenApi\Attributes as OAT;
 #[OAT\Schema()]
 class ShoppingList
 {
-    #[OAT\Property(property: 'ingredient_id', example: 1)]
-    public int $ingredientId;
-    #[OAT\Property(property: 'ingredient_name', example: 'gin-1')]
-    public string $ingredientName;
-    #[OAT\Property(property: 'ingredient_slug', example: 'gin-1')]
-    public string $ingredientSlug;
+    #[OAT\Property()]
+    public IngredientBasic $ingredient;
+    #[OAT\Property(example: 3)]
+    public int $quantity;
 }